requirements.txt
praw==7.8.1
google-cloud-storage==2.18.0
python-dotenv==0.21.0
google-cloud-bigquery==3.26.0
